# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeoBank API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileTier =
  (typeof UserProfileTier)[keyof typeof UserProfileTier];

export const UserProfileTier = {
  standard: "standard",
  premium: "premium",
  elite: "elite",
} as const;

export interface UserProfile {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  avatarUrl?: string;
  memberSince: string;
  tier: UserProfileTier;
  address?: string;
  city?: string;
  country?: string;
  iban?: string;
  bic?: string;
}

export interface DemoLoginResponse {
  success: boolean;
  user: UserProfile;
  message: string;
}

export interface CategorySpend {
  category: string;
  amount: number;
  percentage: number;
  color: string;
}

export interface BalancePoint {
  date: string;
  balance: number;
}

export interface UserStats {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  savingsRate: number;
  spendingByCategory: CategorySpend[];
  balanceHistory: BalancePoint[];
}

export type AccountType = (typeof AccountType)[keyof typeof AccountType];

export const AccountType = {
  checking: "checking",
  savings: "savings",
  business: "business",
  investment: "investment",
} as const;

export interface Account {
  id: string;
  type: AccountType;
  name: string;
  iban: string;
  balance: number;
  currency: string;
  color?: string;
  isDefault: boolean;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  debit: "debit",
  credit: "credit",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  completed: "completed",
  pending: "pending",
  failed: "failed",
} as const;

export interface Transaction {
  id: string;
  date: string;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  merchant?: string;
  status: TransactionStatus;
  icon?: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
}

export type CardType = (typeof CardType)[keyof typeof CardType];

export const CardType = {
  virtual: "virtual",
  physical: "physical",
} as const;

export type CardNetwork = (typeof CardNetwork)[keyof typeof CardNetwork];

export const CardNetwork = {
  visa: "visa",
  mastercard: "mastercard",
} as const;

export interface Card {
  id: string;
  type: CardType;
  network: CardNetwork;
  last4: string;
  cardholderName: string;
  expiryMonth: number;
  expiryYear: number;
  isLocked: boolean;
  creditLimit?: number;
  currentBalance?: number;
  color?: string;
  accountId: string;
}

export interface TransferRequest {
  fromAccountId: string;
  toIban: string;
  toName: string;
  amount: number;
  currency: string;
  description?: string;
}

export interface TransferResponse {
  success: boolean;
  transactionId: string;
  message: string;
}

export type LoanProductType =
  (typeof LoanProductType)[keyof typeof LoanProductType];

export const LoanProductType = {
  personal: "personal",
  auto: "auto",
  mortgage: "mortgage",
  business: "business",
  student: "student",
} as const;

export interface LoanProduct {
  id: string;
  name: string;
  type: LoanProductType;
  minAmount: number;
  maxAmount: number;
  minDurationMonths: number;
  maxDurationMonths: number;
  minRate: number;
  maxRate: number;
  description: string;
  features: string[];
}

export interface LoanSimulationRequest {
  productId: string;
  amount: number;
  durationMonths: number;
}

export interface PaymentScheduleEntry {
  month: number;
  payment: number;
  principal: number;
  interest: number;
  remainingBalance: number;
}

export interface LoanSimulationResult {
  monthlyPayment: number;
  totalRepayment: number;
  totalInterest: number;
  annualRate: number;
  schedule: PaymentScheduleEntry[];
}

export interface LoanApplicationRequest {
  productId: string;
  amount: number;
  durationMonths: number;
  purpose: string;
  employmentType: string;
  monthlyIncome: number;
}

export type LoanApplicationStatus =
  (typeof LoanApplicationStatus)[keyof typeof LoanApplicationStatus];

export const LoanApplicationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  active: "active",
  closed: "closed",
} as const;

export interface LoanApplication {
  id: string;
  productId: string;
  productName: string;
  amount: number;
  durationMonths: number;
  monthlyPayment: number;
  annualRate: number;
  status: LoanApplicationStatus;
  appliedAt: string;
  purpose: string;
}

export type SavingsGoalAutoSaveFrequency =
  (typeof SavingsGoalAutoSaveFrequency)[keyof typeof SavingsGoalAutoSaveFrequency];

export const SavingsGoalAutoSaveFrequency = {
  weekly: "weekly",
  monthly: "monthly",
  none: "none",
} as const;

export interface SavingsGoal {
  id: string;
  name: string;
  targetAmount: number;
  currentAmount: number;
  targetDate: string;
  category: string;
  color: string;
  autoSaveAmount?: number;
  autoSaveFrequency?: SavingsGoalAutoSaveFrequency;
}

export type CreateSavingsGoalRequestAutoSaveFrequency =
  (typeof CreateSavingsGoalRequestAutoSaveFrequency)[keyof typeof CreateSavingsGoalRequestAutoSaveFrequency];

export const CreateSavingsGoalRequestAutoSaveFrequency = {
  weekly: "weekly",
  monthly: "monthly",
  none: "none",
} as const;

export interface CreateSavingsGoalRequest {
  name: string;
  targetAmount: number;
  targetDate: string;
  category: string;
  autoSaveAmount?: number;
  autoSaveFrequency?: CreateSavingsGoalRequestAutoSaveFrequency;
}

export type GetAccountTransactionsParams = {
  limit?: number;
  offset?: number;
};