# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NeoBank API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts/{accountId}/transactions:
    get:
      operationId: getAccountTransactions
      tags: [transactions]
      summary: Get transactions for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
components:
  schemas:
...
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: number
      required: [transactions, total]
    Transaction:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
        category:
          type: string
        merchant:
          type: string
        status:
          type: string
          enum: [completed, pending, failed]
        icon:
          type: string
      required: [id, date, description, amount, type, category, status]
...